--- a/packages/sdk/package.json
+++ b/packages/sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mocksi/bilan-sdk",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "description": "TypeScript SDK for Bilan trust analytics",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",