--- a/packages/server/package.json
+++ b/packages/server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bilan/server",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "description": "Self-hostable server for Bilan trust analytics",
   "main": "dist/index.js",
   "type": "module",