… README.md
… tests/deployment/docker.test.ts
… tests/deployment/health-checks.test.ts
… tests/e2e/deployment.test.ts
… tests/performance/api-benchmarks.test.ts
… tests/performance/dashboard-benchmarks.test.ts